refactor(getinfo): split into _http/_metadata/_query/_stats submodules
Phase 2 of TDD refactor. Splits restgdf/utils/getinfo.py into four private
submodules by concern while preserving all public import paths and patch
seams:

- _http.py: DEFAULTDICT, DEFAULT_METADATA_HEADERS, default_headers, default_data
- _metadata.py: FIELDDOESNOTEXIST, supports_pagination, get_object_id_field,
  get_max_record_count, get_name, getfields, getfields_df
- _query.py: get_feature_count, get_metadata, get_object_ids
- _stats.py: getuniquevalues, getvaluecounts, nestedcount

get_offset_range and service_metadata remain defined in getinfo.py (the
shim) so that patches against restgdf.utils.getinfo.<helper> continue to
intercept their sibling calls via module-namespace resolution. This is
pinned by the new tests/test_getinfo_seams.py regression suite.

All new submodules begin with 'from __future__ import annotations' for
Py3.9 runtime compatibility with PEP 604 type syntax.

162 tests pass (157 prior + 5 new seams); pre-commit green.
