+++ restgdf/utils/_http.py
@@ -1,0 +1,34 @@
+"""HTTP defaults for ArcGIS REST helpers.
+
+Private submodule; all public names are re-exported by
+``restgdf.utils.getinfo`` to preserve import paths.
+"""
+
+from __future__ import annotations
+
+DEFAULT_METADATA_HEADERS = {
+    "Accept": "application/json,text/plain,*/*",
+    "User-Agent": "Mozilla/5.0",
+}
+
+DEFAULTDICT: dict = {
+    "where": "1=1",
+    "outFields": "*",
+    "returnGeometry": True,
+    "returnCountOnly": False,
+    "f": "json",
+}
+
+
+def default_headers(headers: dict | None = None) -> dict:
+    """Return request headers merged with ArcGIS-compatible defaults."""
+    return {**DEFAULT_METADATA_HEADERS, **(headers or {})}
+
+
+def default_data(
+    data: dict | None = None,
+    default_dict: dict | None = None,
+) -> dict:
+    """Return a dict with default values for ArcGIS REST API requests."""
+    default_dict = default_dict or DEFAULTDICT
+    return {**default_dict, **(data or {})}
